scripts/python/municipalidad/requirements.txt
distro==1.5.0
numpy==1.20.3
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
tabula-py==2.2.0
